Refactor frontend: move login logic to standalone JS module and restructure component paths.
