+++ src/main/resources/frontend/index.html
@@ -6,39 +6,7 @@
 </head>
 
 <body>
-<h1>Login</h1>
-<input id="email" placeholder="Email" />
-<input id="password" placeholder="Hasło" type="password" />
-<button onclick="login()">Zaloguj się</button>
-
-<div id="result"></div>
-
-<script>
-    async function login() {
-        const email = document.getElementById('email').value;
-        const password = document.getElementById('password').value;
-
-        const response = await fetch('https://reqres.in/api/login', {
-            method: 'POST',
-            headers: { 'Content-Type': 'application/json' },
-            body: JSON.stringify({ email, password })
-        });
-
-        if (response.ok) {
-            const data = await response.json();
-            document.getElementById('result').innerText = 'Token: ' + data.token;
-            fetchUsers();
-        } else {
-            document.getElementById('result').innerText = 'Błąd logowania';
-        }
-    }
-
-    async function fetchUsers() {
-        const res = await fetch('https://reqres.in/api/users?page=1');
-        const json = await res.json();
-        const users = json.data.map(u => `${u.first_name} ${u.last_name}`).join('<br>');
-        document.getElementById('result').innerHTML += '<br><br><strong>Użytkownicy:</strong><br>' + users;
-    }
+<script src="app.js" type="module" >
 </script>
 </body>
 </html>
