+++ src/main/resources/frontend/login/App.js
@@ -1,5 +1,5 @@
 import React from "react";
-import LoginForm from "../components/LoginForm";
+import LoginForm from "./LoginForm";
 
 function App() {
     return (
